added object tests, split into default, array and object plugin testing, fixed config resolving function
