+++ test/mock/env-manager.js
